+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Date format: (YYYY-MM-DD)
 * Bumped CraftBukkit dependency for MC 1.15 to 1.15.2.
 * Changed: Replaced the NMS call for setting the CanJoinRaid property on MC 1.15.x with a corresponding Bukkit API call. This only works on the very latest versions of MC 1.15.1 and upwards.
 * Added: Warning messages when a trading offer cannot be loaded for some reason.
+* Changed: Trading offer and hire cost items get automatically migrated to the current data version now. This will typically occur with every (even minor) minecraft updates. A warning is logged if an offer could not be migrated.
+* Config/Debug: Added debug option 'item-migrations' to log whenever a shopkeeper performs item migrations.
 
 ## v2.9.0 (2019-12-28)
 ### Supported MC versions: 1.15.1, 1.14.4